backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
pydantic
pydantic-settings
python-dotenv
